Move the ViewType impls into unconditional code.
Move the ViewType impls into code that is used for both the std and
polyfill configurations.
