+++ Cargo.toml
@@ -22,7 +22,7 @@ fs-err = { version = "2.6.0", optional = true }
 
 [target.'cfg(not(target_os = "wasi"))'.dependencies]
 # Optionally depend on os_pipe to implement traits for its types for now.
-os_pipe = { version = "1.0.0", optional = true }
+os_pipe = { version = "1.0.0", features = ["io_safety"], optional = true }
 
 # The following dependencies allow io-lifetimes to define impls for various
 # third-party traits. This is only done in not(io_lifetimes_use_std) mode,
