+++ src/impls_mio.rs
@@ -2,9 +2,6 @@
 //! future, we'll prefer to have crates provide their own impls; this is
 //! just a temporary measure.
 
-#[cfg(unix)]
-use crate::views::FilelikeViewType;
-use crate::views::SocketlikeViewType;
 #[cfg(any(unix, target_os = "wasi"))]
 use crate::{AsFd, BorrowedFd, FromFd, IntoFd, OwnedFd};
 #[cfg(windows)]
@@ -15,8 +12,6 @@ use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd};
 use std::os::wasi::io::{AsRawFd, FromRawFd, IntoRawFd};
 #[cfg(windows)]
 use std::os::windows::io::{AsRawSocket, FromRawSocket, IntoRawSocket};
-
-unsafe impl SocketlikeViewType for mio::net::TcpStream {}
 
 #[cfg(any(unix, target_os = "wasi"))]
 impl AsFd for mio::net::TcpStream {
@@ -98,8 +93,6 @@ impl From<OwnedSocket> for mio::net::TcpStream {
     }
 }
 
-unsafe impl SocketlikeViewType for mio::net::TcpListener {}
-
 #[cfg(any(unix, target_os = "wasi"))]
 impl AsFd for mio::net::TcpListener {
     #[inline]
@@ -179,8 +172,6 @@ impl From<OwnedSocket> for mio::net::TcpListener {
         unsafe { Self::from_raw_socket(owned.into_raw_socket()) }
     }
 }
-
-unsafe impl SocketlikeViewType for mio::net::UdpSocket {}
 
 #[cfg(any(unix, target_os = "wasi"))]
 impl AsFd for mio::net::UdpSocket {
@@ -263,9 +254,6 @@ impl From<OwnedSocket> for mio::net::UdpSocket {
 }
 
 #[cfg(unix)]
-unsafe impl SocketlikeViewType for mio::net::UnixDatagram {}
-
-#[cfg(unix)]
 impl AsFd for mio::net::UnixDatagram {
     #[inline]
     fn as_fd(&self) -> BorrowedFd<'_> {
@@ -304,9 +292,6 @@ impl From<OwnedFd> for mio::net::UnixDatagram {
         unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
     }
 }
-
-#[cfg(unix)]
-unsafe impl SocketlikeViewType for mio::net::UnixListener {}
 
 #[cfg(unix)]
 impl AsFd for mio::net::UnixListener {
@@ -349,9 +334,6 @@ impl From<OwnedFd> for mio::net::UnixListener {
 }
 
 #[cfg(unix)]
-unsafe impl SocketlikeViewType for mio::net::UnixStream {}
-
-#[cfg(unix)]
 impl AsFd for mio::net::UnixStream {
     #[inline]
     fn as_fd(&self) -> BorrowedFd<'_> {
@@ -392,9 +374,6 @@ impl From<OwnedFd> for mio::net::UnixStream {
 }
 
 #[cfg(unix)]
-unsafe impl FilelikeViewType for mio::unix::pipe::Receiver {}
-
-#[cfg(unix)]
 impl AsFd for mio::unix::pipe::Receiver {
     #[inline]
     fn as_fd(&self) -> BorrowedFd<'_> {
@@ -433,9 +412,6 @@ impl From<OwnedFd> for mio::unix::pipe::Receiver {
         unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
     }
 }
-
-#[cfg(unix)]
-unsafe impl FilelikeViewType for mio::unix::pipe::Sender {}
 
 #[cfg(unix)]
 impl AsFd for mio::unix::pipe::Sender {
